Struct inside struct are interpreted as a pointer
Fixes how struct inside struct are handled by hxcpp (in older versions,
they did a copy of the struct on a temporary variable - which of course
wasn't changeable). This also makes it more natural to work with structs
inside structs on languages that don't support structs - like Java

This happens partly because in Haxe, the following must hold true:

```haxe
some.field.access = someval
// should be the same as
var s1 = some.field;
s1.access = someval
```

This doesn't happen when working with naked structs, so in this way,
it's better to restrict working with pointers to structs.